# docs/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":"A simple greeting for this API","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":[""]}},"/crawl":{"post":{"operationId":"crawl","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSnapshotDto"}}}},"responses":{"200":{"description":"Returns snapshot created","content":{"application/json":{"schema":{"type":"object","properties":{"_id":{"type":"string","example":"1"},"url":{"type":"string","example":"https://www.google.com"},"createdAt":{"type":"string","example":"2021-03-01T00:00:00.000Z"},"screenshot":{"type":"string","example":"/tmp/screenshots/1234.png","title":"Screenshot path in the server"},"documentTitle":{"type":"string","example":"Google"},"outgoingUrls":{"type":"object","title":"Outgoing URLs originating from the crawled page","properties":{"document":{"type":"array","items":{"type":"string","example":"https://www.google.com"}},"script":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"stylesheet":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"image":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/images/logo.svg"}},"xhr":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-content/uploads/sites/5/2023/02/homepage-illustration-desktop.json"}},"font":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/fonts/graphikgett/GraphikGett-Semibold.woff2"}},"fetch":{"type":"array","items":{"type":"string","example":"https://js.nagich.co.il/assets/scripts/locale.js"}},"ping":{"type":"array","items":{"type":"string","example":"https://pagead2.googlesyndication.com/pagead/landing?gcs=G100&gcd=G100&rnd=661027870.1677160927&url=https%3A%2F%2Fwww.google.com%2Fil%2F&gtm=1231512413"}}}},"links":{"type":"array","title":"Links (anchor elements) found in the crawled page","items":{"type":"string","examples":["https://www.example.com","https://www.facebook.com"]}},"stylesheets":{"type":"array","title":"Stylesheets referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"scripts":{"type":"array","title":"Scripts referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"metadata":{"type":"object","title":"Metadata","properties":{"title":{"type":"string","example":"My website"},"description":{"type":"string","example":"This is a website description"},"icon":{"type":"string","example":"https://www.my-website.com/favicon.ico"},"image":{"type":"string","example":"https://www.my-website.com/logo.png"},"author":{"type":"string","example":"John Doe"}}}}}}}},"400":{"description":"Bad Request - e.g. url was not found"}},"tags":[""]}},"/snapshots":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSnapshotDto"}}}},"responses":{"200":{"description":"Returns snapshot created","content":{"application/json":{"schema":{"type":"object","properties":{"_id":{"type":"string","example":"1"},"url":{"type":"string","example":"https://www.google.com"},"createdAt":{"type":"string","example":"2021-03-01T00:00:00.000Z"},"screenshot":{"type":"string","example":"/tmp/screenshots/1234.png","title":"Screenshot path in the server"},"documentTitle":{"type":"string","example":"Google"},"outgoingUrls":{"type":"object","title":"Outgoing URLs originating from the crawled page","properties":{"document":{"type":"array","items":{"type":"string","example":"https://www.google.com"}},"script":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"stylesheet":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"image":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/images/logo.svg"}},"xhr":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-content/uploads/sites/5/2023/02/homepage-illustration-desktop.json"}},"font":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/fonts/graphikgett/GraphikGett-Semibold.woff2"}},"fetch":{"type":"array","items":{"type":"string","example":"https://js.nagich.co.il/assets/scripts/locale.js"}},"ping":{"type":"array","items":{"type":"string","example":"https://pagead2.googlesyndication.com/pagead/landing?gcs=G100&gcd=G100&rnd=661027870.1677160927&url=https%3A%2F%2Fwww.google.com%2Fil%2F&gtm=1231512413"}}}},"links":{"type":"array","title":"Links (anchor elements) found in the crawled page","items":{"type":"string","examples":["https://www.example.com","https://www.facebook.com"]}},"stylesheets":{"type":"array","title":"Stylesheets referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"scripts":{"type":"array","title":"Scripts referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"metadata":{"type":"object","title":"Metadata","properties":{"title":{"type":"string","example":"My website"},"description":{"type":"string","example":"This is a website description"},"icon":{"type":"string","example":"https://www.my-website.com/favicon.ico"},"image":{"type":"string","example":"https://www.my-website.com/logo.png"},"author":{"type":"string","example":"John Doe"}}}}}}}},"400":{"description":"Bad Request - e.g. url was not found"}},"tags":["Snapshots"]},"get":{"operationId":"findAll","parameters":[{"name":"orderBy","required":false,"in":"query","description":"[SORT] The field to order by","schema":{"default":"createdAt","enum":["url","createdAt"],"type":"string"}},{"name":"orderDirection","required":false,"in":"query","description":"[SORT] The order direction (Ascending or Descending)","schema":{"default":"DESC","enum":["ASC","DESC"],"type":"string"}},{"name":"url","required":false,"in":"query","description":"[FILTER] Only return snapshots with certain URL","example":"https://google.com","schema":{"default":"None (all URLs)","type":"string"}},{"name":"createdBefore","required":false,"in":"query","description":"[FILTER] Only return snapshots created BEFORE certain date/datetime","example":"2022-08-10, 2023-01-10T10:35:00.000Z","schema":{"default":"None (all URLs)","type":"string"}},{"name":"createdAfter","required":false,"in":"query","description":"[FILTER] Only return snapshots created AFTER certain date/datetime","example":"2022-08-10, 2023-01-10T10:35:00.000Z","schema":{"default":"None (all URLs)","type":"string"}}],"responses":{"200":{"description":"Returns snapshot created","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string","example":"63f75a02a86bcf0aa3803523"},"url":{"type":"string","example":"https://www.google.com"},"createdAt":{"type":"string","example":"2021-03-01T00:00:00.000Z"},"documentTitle":{"type":"string","example":"Google"}}}}}}}},"tags":["Snapshots"]}},"/snapshots/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns snapshot","content":{"application/json":{"schema":{"type":"object","properties":{"_id":{"type":"string","example":"1"},"url":{"type":"string","example":"https://www.google.com"},"createdAt":{"type":"string","example":"2021-03-01T00:00:00.000Z"},"screenshot":{"type":"string","example":"/tmp/screenshots/1234.png","title":"Screenshot path in the server"},"documentTitle":{"type":"string","example":"Google"},"outgoingUrls":{"type":"object","title":"Outgoing URLs originating from the crawled page","properties":{"document":{"type":"array","items":{"type":"string","example":"https://www.google.com"}},"script":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"stylesheet":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"image":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/images/logo.svg"}},"xhr":{"type":"array","items":{"type":"string","example":"https://www.google.com/il/wp-content/uploads/sites/5/2023/02/homepage-illustration-desktop.json"}},"font":{"type":"array","items":{"type":"string","example":"https://www.google.com/wp-content/themes/gett/build/fonts/graphikgett/GraphikGett-Semibold.woff2"}},"fetch":{"type":"array","items":{"type":"string","example":"https://js.nagich.co.il/assets/scripts/locale.js"}},"ping":{"type":"array","items":{"type":"string","example":"https://pagead2.googlesyndication.com/pagead/landing?gcs=G100&gcd=G100&rnd=661027870.1677160927&url=https%3A%2F%2Fwww.google.com%2Fil%2F&gtm=1231512413"}}}},"links":{"type":"array","title":"Links (anchor elements) found in the crawled page","items":{"type":"string","examples":["https://www.example.com","https://www.facebook.com"]}},"stylesheets":{"type":"array","title":"Stylesheets referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/css/dist/block-library/style-rtl.min.css?ver=6.1.1"}},"scripts":{"type":"array","title":"Scripts referenced in the crawled page","items":{"type":"string","example":"https://www.google.com/il/wp-includes/js/jquery/jquery.min.js?ver=3.6.1"}},"metadata":{"type":"object","title":"Metadata","properties":{"title":{"type":"string","example":"My website"},"description":{"type":"string","example":"This is a website description"},"icon":{"type":"string","example":"https://www.my-website.com/favicon.ico"},"image":{"type":"string","example":"https://www.my-website.com/logo.png"},"author":{"type":"string","example":"John Doe"}}}}}}}},"404":{"description":"Snapshot not found with specified id"}},"tags":["Snapshots"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Empty response"},"404":{"description":"Snapshot not found with specified id"}},"tags":["Snapshots"]}}},"info":{"title":"Web Crawler & Snapshots API","description":"An API to web crawling and asset harvesting, and to manage snapshots of web pages.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateSnapshotDto":{"type":"object","properties":{"url":{"type":"string","description":"The URL of the site to be crawled","example":"https://google.com"}},"required":["url"]}}}}